+++ src/Service/BedrockRuntime/src/Input/InvokeModelRequest.php
@@ -48,9 +48,9 @@ final class InvokeModelRequest extends Input
      *   see Supported Regions and models for cross-region inference [^2] in the Amazon Bedrock User Guide.
      * - If you use a provisioned model, specify the ARN of the Provisioned Throughput. For more information, see Run
      *   inference using a Provisioned Throughput [^3] in the Amazon Bedrock User Guide.
-     * - If you use a custom model, first purchase Provisioned Throughput for it. Then specify the ARN of the resulting
-     *   provisioned model. For more information, see Use a custom model in Amazon Bedrock [^4] in the Amazon Bedrock User
-     *   Guide.
+     * - If you use a custom model, specify the ARN of the custom model deployment (for on-demand inference) or the ARN of
+     *   your provisioned model (for Provisioned Throughput). For more information, see Use a custom model in Amazon Bedrock
+     *   [^4] in the Amazon Bedrock User Guide.
      * - If you use an imported model [^5], specify the ARN of the imported model. You can get the model ARN from a
      *   successful call to CreateModelImportJob [^6] or from the Imported models page in the Amazon Bedrock console.
      *