+++ src/Service/CloudFormation/src/ValueObject/RollbackTrigger.php
@@ -24,8 +24,8 @@ final class RollbackTrigger
      * The resource type of the rollback trigger. Specify either AWS::CloudWatch::Alarm [^1] or
      * AWS::CloudWatch::CompositeAlarm [^2] resource types.
      *
-     * [^1]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-cloudwatch-alarm.html
-     * [^2]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-cloudwatch-compositealarm.html
+     * [^1]: https://docs.aws.amazon.com/AWSCloudFormation/latest/TemplateReference/aws-resource-cloudwatch-alarm.html
+     * [^2]: https://docs.aws.amazon.com/AWSCloudFormation/latest/TemplateReference/aws-resource-cloudwatch-compositealarm.html
      *
      * @var string
      */