+++ src/Service/DynamoDb/src/Enum/WitnessStatus.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace AsyncAws\DynamoDb\Enum;
+
+final class WitnessStatus
+{
+    public const ACTIVE = 'ACTIVE';
+    public const CREATING = 'CREATING';
+    public const DELETING = 'DELETING';
+
+    public static function exists(string $value): bool
+    {
+        return isset([
+            self::ACTIVE => true,
+            self::CREATING => true,
+            self::DELETING => true,
+        ][$value]);
+    }
+}